introduce VITE_DEFAULT_DATA_ENV
