+++ packages/atlas/src/.env
@@ -2,11 +2,13 @@
 
 # should be set to "production" for production builds
 VITE_ENV=development
+VITE_ENV_SELECTION_ENABLED=true
+# default env in environments admin modal. Can be production, development, next or local. If not provided, VITE_ENV will be used
+VITE_DEFAULT_DATA_ENV=
 
 # App configuration
 VITE_APP_ID=4414-2
 VITE_APP_NAME=Atlas
-VITE_ENV_SELECTION_ENABLED=true
 
 VITE_AVATAR_SERVICE_URL=https://atlas-services.joystream.org/avatars
 VITE_GEOLOCATION_SERVICE_URL=https://geolocation.joystream.org
