+++ packages/atlas/src/providers/environment/store.ts
@@ -1,19 +1,22 @@
+import { getEnvName } from '@/utils/envVariables'
 import { createStore } from '@/utils/store'
 
 const LOCAL_STORAGE_KEY = 'environment'
 
 export type EnvironmentState = {
-  targetDevEnv: string
+  defaultDataEnv: string
   nodeOverride: string | null
 }
 
+export const ENV_PREFIX = 'VITE'
+
 const INITIAL_STATE: EnvironmentState = {
-  targetDevEnv: 'development',
+  defaultDataEnv: import.meta.env[getEnvName('DEFAULT_DATA_ENV')] || import.meta.env[getEnvName('ENV')] || 'production',
   nodeOverride: null,
 }
 
 export type EnvironmentStoreActions = {
-  setTargetDevEnv: (env: string) => void
+  setDefaultDataEnv: (env: string) => void
   setNodeOverride: (node: string | null) => void
   reset: () => void
   getInitialState: () => EnvironmentState
@@ -28,9 +31,9 @@ export const useEnvironmentStore = createStore<EnvironmentState, EnvironmentStor
           state.nodeOverride = node
         })
       },
-      setTargetDevEnv: (env) => {
+      setDefaultDataEnv: (env) => {
         set((state) => {
-          state.targetDevEnv = env
+          state.defaultDataEnv = env
         })
       },
       reset: () => {
@@ -43,7 +46,7 @@ export const useEnvironmentStore = createStore<EnvironmentState, EnvironmentStor
     persist: {
       key: LOCAL_STORAGE_KEY,
       version: 0,
-      whitelist: ['nodeOverride', 'targetDevEnv'],
+      whitelist: ['nodeOverride', 'defaultDataEnv'],
       migrate: () => null,
     },
   }
